Add BayesWBase
This allows us to share code between the sparse and dense
implementations.

The sparse implementation looks like it is running correctly.

The dense implementation looks like it is estimating heritability a bit
too high.
