+++ src/densebayesw.h
@@ -8,118 +8,41 @@
 #ifndef DENSEBAYESW_HPP_
 #define DENSEBAYESW_HPP_
 
-#include "bayeswcommon.h"
+#include "bayeswbase.h"
 #include "data.hpp"
 #include "options.hpp"
 #include "distributions_boost.hpp"
 
-#include <Eigen/Eigen>
+struct dense_gh_params : public gh_params {
+    dense_gh_params(const VectorXd &vi, const VectorXd &Xj) : vi(vi), Xj(Xj) {}
+    const VectorXd &vi;
+    const VectorXd &Xj;
 
-#if 0
-// Two structures for ARS
-struct pars{
-	/* Common parameters for the densities */
-	VectorXd epsilon;			// epsilon per subject (before each sampling, need to remove the effect of the sampled parameter and then carry on
-
-	VectorXd mixture_classes; // Vector to store mixture component C_k values
-
-	int used_mixture; //Write the index of the mixture we decide to use
-
-	/* Store the current variables */
-	double alpha, sigma_b;
-
-	/* Beta_j - specific variables */
-	VectorXd X_j;
-	/* Mu-specific variables */
-	double sigma_mu;
-	/* sigma_b-specific variables */
-	double alpha_sigma, beta_sigma;
-
-	/*  of sum(X_j*failure) */
-	double sum_failure;
-
-	/* Number of events (sum of failure indicators) */
-	double d;
-
-	/* Help variable for storing sqrt(2sigma_b)	 */
-	double sqrt_2sigmab;
-
+    double exponent_sum() const override;
+    double integrand_adaptive(double s, double alpha, double dj, double sqrt_2Ck_sigmab) const override;
 };
 
-struct pars_alpha{
-	VectorXd failure_vector;
-	VectorXd epsilon;			// epsilon per subject (before each sampling, need to remove the effect of the sampled parameter and then carry on
-
-	/* Alpha-specific variables */
-	double alpha_0, kappa_0;  /*  Prior parameters */
-
-	/* Number of events (sum of failure indicators) */
-	double d;
-};
-
-#endif
-
-
-class DenseBayesW
+class DenseBayesW : public BayesWBase
 {
-	Data            &data; // data matrices
-	Options         &opt;
-	const string    bedFile; // bed file
-	const long      memPageSize; // size of memory
-	const string    outputFile;
-	const int       seed;
-	const int       max_iterations;
-	const int		burn_in;
-	const int       thinning;
-	const double	alpha_0  = 0.01;
-	const double	kappa_0     = 0.01;
-	const double    sigma_mu    = 100;
-	const double    alpha_sigma  = 1;
-	const double    beta_sigma   = 0.0001;
-	const string 	quad_points = opt.quad_points;  // Number of Gaussian quadrature points
-	const int 		K = opt.S.size()+1;  //number of mixtures + 0 class
-
-	Distributions_boost dist;
-
-	// The ARS variables
-	struct pars used_data;
-	struct pars_alpha used_data_alpha;
-
-	// Component variables
-	VectorXd pi_L;        // mixture probabilities
-	VectorXd marginal_likelihoods;      // likelihood for each mixture component
-	VectorXd v;         // variable storing the component assignment
-	VectorXi components; // Indicator vector stating to which mixture SNP belongs to
-
-	// Linear model variables
-	VectorXd theta;		 // Fixed effect sizes
-	VectorXd beta;       // effect sizes
-	VectorXd vi;		 // adjusted and exponented epsilon
-
-	VectorXd y;
-	VectorXd sum_failure;
-	VectorXd sum_failure_fix;
-
-	//Sampled variables (not kept in struct)
-	double mu;
-
-
-
 public:
-        DenseBayesW(Data &data, Options &opt, const long memPageSize);
-        virtual ~DenseBayesW();
-	int runGibbs_Gauss(); // where we run Gibbs sampling over the parametrised model
+    DenseBayesW(Data &data, Options &opt, const long memPageSize);
 
+protected:
+    void sampleBeta(int marker) override;
 
-private:
-	void init(unsigned int markerCount, unsigned int individualCount, unsigned int fixedCount);
-	void sampleMu();
-	void sampleTheta(int fix_i);
-	void sampleBeta(int marker);
-	void sampleAlpha();
+    double calculateSumFailure(int marker) override;
 
-	void marginal_likelihood_vec_calc(VectorXd prior_prob, VectorXd &post_marginals, VectorXd vi, string n);
-	double gauss_hermite_adaptive_integral(int k, VectorXd vi, double sigma, string n);
+    void preEstimateResidualUpdate(int marker) override;
+
+    std::unique_ptr<gh_params> gaussHermiteParameters(int marker) override;
+
+    int estimateBeta (double *xinit, int ninit, double *xl, double *xr,
+                      double *convex, int npoint, int dometrop, double *xprev, double *xsamp,
+                      int nsamp, double *qcent, double *xcent,
+                      int ncent, int *neval) override;
+
+    void postEstimateResidualUpdate(int marker) override;
+
 };
